Rename "activate" to "stimulate" and refactor.
Renamed "activate" methods in the interaction and drawing sections to "stimulate," since "stimulate" is an action that can happen many times. "Activate" implies an initialization action, so the user would expect it to run in p5's setup function rather than the draw loop.

Removed global variable containing the graph windows since it was no longer needed for the object activation. Replaced it with a global variable for the press history, which was previously attached to each graph window (it might still make sense to implement it at the graph-window level, but a global variable is simpler for now).

Removed unnecessary activation code in runMathemagicalListeners and renamed it to updatePressHistory, since that's all it does now.

Removed unnecessary (and problematic) code in the activate method of the interaction objects (the only interaction object currently implemented is the Draggable). The code I removed registered a [drawing object, interaction object] pair with a graph window by storing the pair in an array that I also removed; the registration was happening in every frame of the draw loop.

Deleted comment in the update method of the Rotation animation object (the only animation object currently in the prototype) that was no longer relevant, since it related to the code I removed from the "activate" method.