Rename "update" methods as "takeUpdate" and "giveUpdate."
Drawing objects and animation objects each have update methods. In both cases, they both involve updates, but in different directions. The new names reflect this. 

In particular, mySquare.takeUpdate(myRotation) tells mySquare to take update instructions from myRotation, whereas myRotation.giveUpdate(mySquare) tells myRotation to give update instructions to mySquare. 

Most likely, only takeUpdate() needs to be exposed to the user, but there is another advantage to this naming scheme: it's easier to use in other circumstances. For example, we can also have "takeInput" and "giveInput" methods for user interactions.

This solves the problem of finding a satisfactory name for the method that lets drawing objects respond to user interaction.