Refactor Draggable and rename methods.
The listener methods were doing two things: setting a property and then returning it. Some of those properties were used in the implementation of the mousePulled method, but there's a simpler solution that doesn't require them.

So, I removed the unnecessary properties and the code that set them, and I implemented the simpler version of mousePulled.

Also, mousePulled actually just detected whether the mouse was held down, so I renamed it to mouseHeld.

I adjusted the rest of the code accordingly.