Renamed some of the listeners.
Replaced "mouseIsUnpressed" and "mouseUnpressed" with "mouseIsLetGo" and "mouseLetGo," respectively. "Released" is avoided to prevent conflicts with p5, and "unpressed" is unclear (since p5's mouseIsPressed is true as long as the mouse is pressed, mouseIsUnpressed might be interpreted as being true as long as the mouse is not pressed, but it's actually only true when the mouse is first let go).