Add comments to indicate use of native p5 functions
For now at least, I thought it might be helpful to indicate functions that are provided by p5 when we're using them, so that anyone looking at the code immediately knows where to look to find documentation for those functions (as opposed to having to search for their definitions in the Mathemagical.js repo first). This also provides one easy way of determining which p5 functions we're using with a CTRL+F command (although that assumes the commenting strategy is consistently applied).