Improve event listener/handler interface
The previous interface caused confusion. For example, setObjectIsHovered might be interpreted as a setter for a Boolean, rather than a setter for a method that returns a Boolean.

CRITERIA FOR NEW INTERFACE
The interface in this PR is intended to alleviate confusion by satisfying the following design criteria:

* flexible enough to handle a wide range of events
* intuitive enough for beginners 
* consistent enough with p5

NEW INTERFACE
setEventDetector(type, detector)
setEventResponder(type, responder)

The type parameter accepts strings that match p5's names when they exist (e.g. ‘mousepressed’ and ‘mousereleased’) and they typically match Web API names when they do not (e.g. ‘mouseover’ and ‘mouseout’). We may create our own names in some circumstances.

The detector and responder parameters are callback functions.

This PR also renames some internal variables using a more intuitive naming scheme that matches the new interface.

OTHER ADVANTAGES
This approach also has a couple other advantages over the previous approach:

* The names make it clear these are setter methods.
* The type parameter supports the exact names of p5 handlers.

COMPARISON TO OTHER INTERFACES
The approach in this PR is similar to the addEventListener() method of the EventTarget Web API and the .on() method of Konva.js. One major difference is that .setEventDetector() and .setEventHandler() are attached to an interaction object, rather than an event target. This maintains the original design that predates this pull request.

This way, the user only needs to create a single interaction object in order to apply multiple listeners and handlers (e.g. for a draggable square, multiple listeners and handlers are used to change the cursor on mouseover and mouseout events, and also to change the position on drag events).

TESTING
I haven't had time to test this code yet.