Rename enumeration of event types and update comment
Replace the object name "Interaction" with the name "EVENT_TYPES," so that it's more descriptive and so that it follows typical casing conventions for JavaScript. 

Upper snake case isn't always used for objects declared with const since they're not immutable like primitives, and even frozen objects can be changed if they have a property that stores another object, but the event type properties are all immutable strings.

Typically, we can use lower camel case for objects and upper camel case for classes, but "EVENT_TYPES" suggests both immutability and that it's some kind of object (since the name is plural).