Remove support for multiple detectors, enhance square, refactor
MAJOR CHANGES

* Remove multiple detectors:
Remove support for multiple detectors. Having multiple detectors for a single event type turned out to be problematic. For example, suppose the user tries to add a detector that expands the region used to detect mouse-over events. If the mouse is inside the new region and outside the old one, mouse-over and mouse-out events will both fire.

* Enhance square:
Update square drawing object to support stroke-weight property and methods. This is useful for testing addEventResponder().

* Refactor:
Separate responders into two maps (one for default responders, one for user-provided responders). This makes it possible to ensure the defaults always run first and lays the foundation for methods that allow the user to deactivate and reactivate defaults.

MINOR FIXES

* Replace the "in" operator with the has() method, since the has() method is used with Map objects.

* Replace remove() with delete() and rename removeEventResponder to deleteEventResponder (delete() is the Set object method that removes elements).

* Make some changes for consistency.

* Adjust semicolons, whitespace, and comments.