Add verification file for Google
This file is used to demonstrate ownership of the Mathemagical.js repo to Google Search Console, which should help with getting it indexed so that people can find it from a Google web search. Removing this file, even after verification, will cause a loss of access to this property from within Google Search Console (after a grace period).